moves autopilot package to its own module
