+++ agent/consul/state/autopilot_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/hashicorp/consul/agent/consul/autopilot"
+	"github.com/hashicorp/consul/autopilot"
 	"github.com/pascaldekloe/goe/verify"
 )
