+++ agent/operator_endpoint_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/hashicorp/consul/testrpc"
 	"github.com/stretchr/testify/require"
 
-	"github.com/hashicorp/consul/agent/consul/autopilot"
 	"github.com/hashicorp/consul/agent/structs"
 	"github.com/hashicorp/consul/api"
+	"github.com/hashicorp/consul/autopilot"
 	"github.com/hashicorp/consul/sdk/testutil/retry"
 )
