+++ autopilot/go.mod
@@ -1,0 +1,12 @@
+module github.com/hashicorp/consul/autopilot
+
+go 1.13
+
+require (
+	github.com/hashicorp/go-hclog v0.9.1
+	github.com/hashicorp/go-version v1.2.0
+	github.com/hashicorp/raft v1.1.2
+	github.com/hashicorp/serf v0.8.5
+	github.com/pascaldekloe/goe v0.1.0
+	github.com/stretchr/testify v1.5.1
+)
