+++ autopilot/promotion.go
