+++ autopilot/promotion_test.go
