+++ autopilot/structs.go
